Export Sequence Fuzz Models
