+++ packages/dds/sequence/package.json
@@ -52,6 +52,18 @@
 				"types": "./dist/intervalCollection.d.ts",
 				"default": "./dist/intervalCollection.js"
 			}
+		},
+		"./internal/test": {
+			"allow-ff-test-exports": {
+				"import": {
+					"types": "./lib/test/index.d.ts",
+					"default": "./lib/test/index.js"
+				},
+				"require": {
+					"types": "./dist/test/index.d.ts",
+					"default": "./dist/test/index.js"
+				}
+			}
 		}
 	},
 	"main": "lib/index.js",
