+++ packages/dds/sequence/src/test/fuzz/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export { baseIntervalModel, baseSharedStringModel } from "./fuzzUtils.js";
